feat(bot): accept delta input for rank commands

+++ bathbot/src/embeds/osu/rank_score.rs
@@ -20,7 +20,7 @@ pub struct RankRankedScoreEmbed {
 impl RankRankedScoreEmbed {
     pub fn new(
         user: &RedisData<User>,
-        rank: usize,
+        rank: u32,
         rank_holder: UserCompact,
         respektive_user: Option<RespektiveUser>,
     ) -> Self {
